Cleaning code, add dto request in controller product and category, fix bug setpublished product

+++ src/main/java/team/kucing/anabulshopcare/repository/ProductRepository.java
@@ -16,7 +16,7 @@ public interface ProductRepository extends JpaRepository<Product, UUID> {
 
     Page<Product> findByLocation(String location, Pageable pageable);
 
-    Page<Product> findByPriceBetween(BigDecimal startPrice, BigDecimal endPrice, Pageable pageable);
+    Page<Product> findByPriceBetween(double startPrice, double endPrice, Pageable pageable);
 
     Page<Product> findByIsPublished(boolean status, Pageable pageable);
 }
